refactor(api): move architectural primitives to core package
