+++ backend/api/src/main/java/com/anibalxyz/core/domain/error/DomainErrorReason.java
@@ -1,0 +1,3 @@
+package com.anibalxyz.core.domain.error;
+
+public interface DomainErrorReason {}
