+++ backend/api/src/main/java/com/anibalxyz/core/domain/error/EntityNotFoundError.java
@@ -1,0 +1,3 @@
+package com.anibalxyz.core.domain.error;
+
+public interface EntityNotFoundError extends DomainError {}
