+++ backend/api/src/main/java/com/anibalxyz/core/domain/error/InvalidValueError.java
@@ -1,3 +1,3 @@
-package com.anibalxyz.features.common.domain.error;
+package com.anibalxyz.core.domain.error;
 
 public interface InvalidValueError extends DomainError {}
