+++ backend/api/src/main/java/com/anibalxyz/features/auth/api/AuthErrorMapper.java
@@ -2,6 +2,8 @@ package com.anibalxyz.features.auth.api;
 
 import static net.logstash.logback.argument.StructuredArguments.kv;
 
+import com.anibalxyz.core.domain.error.DomainError;
+import com.anibalxyz.core.domain.error.InvalidValueError;
 import com.anibalxyz.features.auth.api.out.AuthErrorCode;
 import com.anibalxyz.features.auth.application.AuthService;
 import com.anibalxyz.features.auth.application.JwtService;
@@ -12,8 +14,6 @@ import com.anibalxyz.features.common.api.ValidationErrorMapper;
 import com.anibalxyz.features.common.api.out.code.CommonErrorCode;
 import com.anibalxyz.features.common.api.out.response.error.ErrorDetail;
 import com.anibalxyz.features.common.api.out.response.error.ErrorResponse;
-import com.anibalxyz.features.common.domain.error.DomainError;
-import com.anibalxyz.features.common.domain.error.InvalidValueError;
 import com.anibalxyz.server.api.ErrorMapper;
 import com.anibalxyz.server.api.ErrorResult;
 import com.anibalxyz.server.api.FeatureErrorMapper;
