+++ backend/api/src/main/java/com/anibalxyz/features/auth/api/JwtMiddleware.java
@@ -1,9 +1,9 @@
 package com.anibalxyz.features.auth.api;
 
+import com.anibalxyz.core.Result;
+import com.anibalxyz.core.application.exception.FailureSignal;
 import com.anibalxyz.features.auth.application.JwtService;
-import com.anibalxyz.features.common.Result;
 import com.anibalxyz.features.common.api.Role;
-import com.anibalxyz.features.common.application.exception.FailureSignal;
 import com.anibalxyz.server.config.modules.startup.StartupConfig;
 import com.anibalxyz.server.context.RequestContext;
 import io.javalin.config.JavalinConfig;
