+++ backend/api/src/main/java/com/anibalxyz/features/auth/domain/error/AuthDomainError.java
@@ -1,10 +1,10 @@
 package com.anibalxyz.features.auth.domain.error;
 
-import com.anibalxyz.features.common.domain.error.DomainError;
+import com.anibalxyz.core.domain.error.DomainError;
 
 public sealed interface AuthDomainError extends DomainError
     permits AuthDomainError.InvalidValueError, InvalidCredentialsError {
   sealed interface InvalidValueError
-      extends com.anibalxyz.features.common.domain.error.InvalidValueError, AuthDomainError
+      extends com.anibalxyz.core.domain.error.InvalidValueError, AuthDomainError
       permits InvalidRefreshTokenError {}
 }
