+++ backend/api/src/main/java/com/anibalxyz/features/common/api/ValidationErrorMapper.java
@@ -1,10 +1,10 @@
 package com.anibalxyz.features.common.api;
 
+import com.anibalxyz.core.application.ValidationNotification;
+import com.anibalxyz.core.domain.error.DomainError;
 import com.anibalxyz.features.common.api.out.code.ValidationErrorCode;
 import com.anibalxyz.features.common.api.out.response.error.ErrorDetail;
 import com.anibalxyz.features.common.api.out.response.error.ErrorResponse;
-import com.anibalxyz.features.common.application.ValidationNotification;
-import com.anibalxyz.features.common.domain.error.DomainError;
 import com.anibalxyz.server.api.ErrorResult;
 import java.util.List;
 import java.util.function.Function;
