--- backend/api/src/main/java/com/anibalxyz/features/common/domain/error/DomainErrorReason.java
@@ -1,3 +1,0 @@
-package com.anibalxyz.features.common.domain.error;
-
-public interface DomainErrorReason {}
