+++ backend/api/src/main/java/com/anibalxyz/features/users/api/UserController.java
@@ -1,9 +1,9 @@
 package com.anibalxyz.features.users.api;
 
-import com.anibalxyz.features.common.Result;
+import com.anibalxyz.core.Result;
+import com.anibalxyz.core.application.ValidationNotification;
+import com.anibalxyz.core.application.exception.FailureSignal;
 import com.anibalxyz.features.common.api.out.response.success.CollectionResponse;
-import com.anibalxyz.features.common.application.ValidationNotification;
-import com.anibalxyz.features.common.application.exception.FailureSignal;
 import com.anibalxyz.features.users.api.in.UserCreateRequest;
 import com.anibalxyz.features.users.api.in.UserUpdateRequest;
 import com.anibalxyz.features.users.api.out.UserDetailResponse;
