+++ backend/api/src/main/java/com/anibalxyz/features/users/api/UserErrorMapper.java
@@ -2,13 +2,13 @@ package com.anibalxyz.features.users.api;
 
 import static net.logstash.logback.argument.StructuredArguments.kv;
 
+import com.anibalxyz.core.domain.error.DomainError;
+import com.anibalxyz.core.domain.error.InvalidValueError;
 import com.anibalxyz.features.common.api.ValidationErrorMapper;
 import com.anibalxyz.features.common.api.out.code.CommonErrorCode;
 import com.anibalxyz.features.common.api.out.code.ValidationErrorCode;
 import com.anibalxyz.features.common.api.out.response.error.ErrorDetail;
 import com.anibalxyz.features.common.api.out.response.error.ErrorResponse;
-import com.anibalxyz.features.common.domain.error.DomainError;
-import com.anibalxyz.features.common.domain.error.InvalidValueError;
 import com.anibalxyz.features.users.application.UserService;
 import com.anibalxyz.features.users.domain.error.*;
 import com.anibalxyz.server.api.ErrorResult;
