+++ backend/api/src/main/java/com/anibalxyz/features/users/domain/PasswordHash.java
@@ -1,7 +1,7 @@
 package com.anibalxyz.features.users.domain;
 
 import com.anibalxyz.annotation.ExcludeFromJacocoGenerated;
-import com.anibalxyz.features.common.Result;
+import com.anibalxyz.core.Result;
 import com.anibalxyz.features.users.domain.error.InvalidPasswordError;
 import com.anibalxyz.features.users.domain.error.InvalidPasswordHashError;
 import java.util.Objects;
