+++ backend/api/src/main/java/com/anibalxyz/server/api/ErrorMapper.java
@@ -3,8 +3,8 @@ package com.anibalxyz.server.api;
 import com.anibalxyz.features.auth.api.AuthErrorMapper;
 import com.anibalxyz.features.common.api.ValidationErrorMapper;
 import com.anibalxyz.features.common.api.out.response.error.ErrorDetail;
-import com.anibalxyz.features.common.application.ValidationNotification;
-import com.anibalxyz.features.common.domain.error.DomainError;
+import com.anibalxyz.core.application.ValidationNotification;
+import com.anibalxyz.core.domain.error.DomainError;
 import com.anibalxyz.features.users.api.UserErrorMapper;
 import com.anibalxyz.server.exception.UnregisteredMapperException;
 import java.util.List;
