+++ backend/api/src/test/java/com/anibalxyz/features/auth/api/AuthControllerTest.java
@@ -16,8 +16,8 @@ import com.anibalxyz.features.auth.application.out.AuthResult;
 import com.anibalxyz.features.auth.domain.RefreshToken;
 import com.anibalxyz.features.auth.domain.error.InvalidCredentialsError;
 import com.anibalxyz.features.auth.domain.error.InvalidRefreshTokenError;
-import com.anibalxyz.features.common.Result;
-import com.anibalxyz.features.common.application.exception.FailureSignal;
+import com.anibalxyz.core.Result;
+import com.anibalxyz.core.application.exception.FailureSignal;
 import com.anibalxyz.shared.Constants;
 import io.javalin.http.Context;
 import io.javalin.http.Cookie;
